# Specs/001-conversational-order-search/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FastService Order Search API (REST Companion)
  description: |
    REST API companion endpoints for the FastService MCP Server.
    The primary interface is MCP (Model Context Protocol), but these REST endpoints
    provide direct access for debugging and non-MCP clients.
  version: 1.0.0
  contact:
    name: FastService Team
servers:
  - url: https://localhost:7001/api
    description: Local development
  - url: https://fastservice-api.azurewebsites.net/api
    description: Production
paths:
...
  /orders/{orderNumber}:
    get:
      operationId: getOrderDetails
      summary: Get full details for a specific order
      tags:
        - Orders
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: integer
          description: The order/repair ID
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetails'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    OrderDetails:
      type: object
      properties:
        orderNumber:
          type: integer
        createdOn:
          type: string
          format: date-time
        modifiedOn:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
          nullable: true
        statusId:
          type: integer
        status:
          type: string
        statusDescription:
          type: string
        statusCategory:
          type: string
        customer:
          $ref: '#/components/schemas/CustomerInfo'
        device:
          $ref: '#/components/schemas/DeviceInfo'
        repair:
          $ref: '#/components/schemas/RepairInfo'
        employee:
          $ref: '#/components/schemas/UserInfo'
        technician:
          $ref: '#/components/schemas/UserInfo'
      required:
        - orderNumber
        - createdOn
        - modifiedOn
        - statusId
        - status
        - customer
        - device
        - employee
        - technician
    CustomerInfo:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        dni:
          type: integer
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
      required:
        - id
        - firstName
        - lastName
        - fullName
        - address
    DeviceInfo:
      type: object
      properties:
        type:
          type: string
        brand:
          type: string
        model:
          type: string
          nullable: true
        serialNumber:
          type: string
          nullable: true
      required:
        - type
        - brand
    RepairInfo:
      type: object
      properties:
        isWarranty:
          type: boolean
        isHomeService:
          type: boolean
        quote:
          type: number
          format: decimal
          nullable: true
        quoteDate:
          type: string
          format: date-time
          nullable: true
        finalPrice:
          type: number
          format: decimal
          nullable: true
        description:
          type: string
          nullable: true
        accessories:
          type: string
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
      required:
        - isWarranty
        - isHomeService
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
      required:
        - id
        - firstName
        - lastName
        - fullName
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
      required:
        - error
        - message